% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BDM}
\alias{BDM}
\title{BDM}
\usage{
BDM(graph)
}
\arguments{
\item{graph}{A numeric matrix representing the graph adjacency matrix}
}
\value{
A numeric vector containing the backward degree metric for each node
}
\description{
this function computes the backward degree metric of a graph
}
