\name{SA24204163-package}
\alias{SA24204163-package}
\docType{package}
\title{
  change point detection based on complex networks and visibility graphs
}
\description{
  we firstly utilize the Visibility graph algorithm proposed by Lacasa et al. to transform a time series into a complex network, and then use the Backward Degree method proposed by Charakopoulos et al. to identify the change points of the time series.
}
\author{
KaiLi Wang

Maintainer: KaiLi Wang <3235399140@qq.com>
}
\keyword{ package }
