% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint-detection.R
\name{visibility}
\alias{visibility}
\title{visibility}
\usage{
visibility(data, m)
}
\arguments{
\item{data}{the time series after diff}

\item{m}{the size of the window}
}
\value{
a graph
}
\description{
this function can transform time series into a complex network using sliding visibility graph algorithm
}
\examples{
\dontrun{
    x<-rnorm(100,0,1)
    visibility(x,100)
}
}
